requirements.txt
asgiref==3.5.2
astroid==2.11.7
black==22.6.0
click==8.1.3
colorama==0.4.5
dill==0.3.5.1
Django==4.0.6
djangorestframework==3.13.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pylint-plugin-utils==0.7
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.1
tzdata==2022.1
wrapt==1.14.1
